Fix phpstan, qa and tests
