+++ composer.json
@@ -52,7 +52,6 @@
   },
   "require-dev": {
     "phpstan/phpstan": "^0.12.98",
-    "phpstan/phpstan-nette": "^0.12.21",
     "symfony/console": "^5.2",
     "symfony/http-client": "^5.3",
     "php-http/httplug": "^2.2",
