+++ phpstan.neon
@@ -7,6 +7,3 @@ parameters:
     fileExtensions:
     	- php
 
-includes:
-	- vendor/phpstan/phpstan-nette/extension.neon
-	- vendor/phpstan/phpstan-nette/rules.neon
