+++ src/DispatcherInterface.php
@@ -1,8 +1,9 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 
 namespace TomasKulhanek\CzechDataBox;
-
 
 use Psr\Http\Message\ResponseInterface;
