+++ tests/Unit/DispatcherTest.php
@@ -18,6 +18,7 @@ use TomasKulhanek\CzechDataBox\Dispatcher;
 
 class DispatcherTest extends TestCase
 {
+    use GeneratePkcs;
     private const TEST_PASS_PHRASE = 'isds';
 
     public function testInfoServices(): void
@@ -76,7 +77,10 @@ class DispatcherTest extends TestCase
             ->with($requestMock);
 
         $dispatcher = new Dispatcher(
-            $clientInterface, $requestFactory, $streamFactory, $uriFactory
+            $clientInterface,
+            $requestFactory,
+            $streamFactory,
+            $uriFactory
         );
         $dispatcher->dispatch($account, Dispatcher::INFO, $xmlBody);
     }
@@ -137,7 +141,10 @@ class DispatcherTest extends TestCase
             ->with($requestMock);
 
         $dispatcher = new Dispatcher(
-            $clientInterface, $requestFactory, $streamFactory, $uriFactory
+            $clientInterface,
+            $requestFactory,
+            $streamFactory,
+            $uriFactory
         );
         $dispatcher->dispatch($account, Dispatcher::SUPPLEMENTARY, $xmlBody);
     }
@@ -198,7 +205,10 @@ class DispatcherTest extends TestCase
             ->with($requestMock);
 
         $dispatcher = new Dispatcher(
-            $clientInterface, $requestFactory, $streamFactory, $uriFactory
+            $clientInterface,
+            $requestFactory,
+            $streamFactory,
+            $uriFactory
         );
         $dispatcher->dispatch($account, Dispatcher::ACCESS, $xmlBody);
     }
@@ -259,7 +269,10 @@ class DispatcherTest extends TestCase
             ->with($requestMock);
 
         $dispatcher = new Dispatcher(
-            $clientInterface, $requestFactory, $streamFactory, $uriFactory
+            $clientInterface,
+            $requestFactory,
+            $streamFactory,
+            $uriFactory
         );
         $dispatcher->dispatch($account, Dispatcher::SEARCH, $xmlBody);
     }
@@ -320,7 +333,10 @@ class DispatcherTest extends TestCase
             ->with($requestMock);
 
         $dispatcher = new Dispatcher(
-            $clientInterface, $requestFactory, $streamFactory, $uriFactory
+            $clientInterface,
+            $requestFactory,
+            $streamFactory,
+            $uriFactory
         );
         $dispatcher->dispatch($account, Dispatcher::OPERATIONS, $xmlBody);
     }
@@ -380,7 +396,10 @@ class DispatcherTest extends TestCase
             ->with($requestMock);
 
         $dispatcher = new Dispatcher(
-            $clientInterface, $requestFactory, $streamFactory, $uriFactory
+            $clientInterface,
+            $requestFactory,
+            $streamFactory,
+            $uriFactory
         );
         $dispatcher->dispatch($account, Dispatcher::OPERATIONS, $xmlBody);
     }
@@ -441,7 +460,10 @@ class DispatcherTest extends TestCase
             ->with($requestMock);
 
         $dispatcher = new Dispatcher(
-            $clientInterface, $requestFactory, $streamFactory, $uriFactory
+            $clientInterface,
+            $requestFactory,
+            $streamFactory,
+            $uriFactory
         );
         $dispatcher->dispatch($account, Dispatcher::OPERATIONS, $xmlBody);
     }
@@ -503,7 +525,10 @@ class DispatcherTest extends TestCase
             ->with($requestMock);
 
         $dispatcher = new Dispatcher(
-            $clientInterface, $requestFactory, $streamFactory, $uriFactory
+            $clientInterface,
+            $requestFactory,
+            $streamFactory,
+            $uriFactory
         );
         $dispatcher->dispatch($account, Dispatcher::OPERATIONS, $xmlBody);
     }
@@ -520,27 +545,6 @@ class DispatcherTest extends TestCase
         $account = new Account();
         $account->setPkcs12Certificate($pkcsContent, $passPhrase);
         return $account;
-    }
-
-    private function generateP12Certificate(string $passPhrase): string
-    {
-        $Info = [
-            "countryName" => "CZ",
-            "stateOrProvinceName" => "Prague",
-            "localityName" => "Prague",
-            "organizationName" => "Tomáš Kulhánek",
-            "organizationalUnitName" => "Test Department",
-            "commonName" => "Tester",
-            "emailAddress" => "jsem+tests@tomaskulhanek.cz",
-        ];
-
-        $Private_Key = null;
-        $Unsigned_Cert = openssl_csr_new($Info, $Private_Key);
-
-        $Signed_Cert = openssl_csr_sign($Unsigned_Cert, null, $Private_Key, 365);
-
-        openssl_pkcs12_export_to_file($Signed_Cert, "test.p12", $Private_Key, $passPhrase);
-        return file_get_contents(__DIR__ . "/../_data/test.p12");
     }
 
     /**
